Remove old FilterObject class
