+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/repository/database/ChatDatabase.kt
@@ -11,7 +11,6 @@ import io.getstream.chat.android.livedata.repository.database.converter.ExtraDat
 import io.getstream.chat.android.livedata.repository.database.converter.FilterObjectConverter
 import io.getstream.chat.android.livedata.repository.database.converter.ListConverter
 import io.getstream.chat.android.livedata.repository.database.converter.MapConverter
-import io.getstream.chat.android.livedata.repository.database.converter.NewFilterObjectConverter
 import io.getstream.chat.android.livedata.repository.database.converter.SetConverter
 import io.getstream.chat.android.livedata.repository.database.converter.SyncStatusConverter
 import io.getstream.chat.android.livedata.repository.domain.channel.ChannelDao
@@ -45,13 +44,12 @@ import io.getstream.chat.android.livedata.repository.domain.user.UserEntity
         CommandInnerEntity::class,
         SyncStateEntity::class
     ],
-    version = 38,
+    version = 39,
     exportSchema = false
 )
 
 @TypeConverters(
     FilterObjectConverter::class,
-    NewFilterObjectConverter::class,
     ExtraDataConverter::class,
     ListConverter::class,
     MapConverter::class,
